Android Pixel Perfect Mode: eliminate conversions between dpi and pixels in Blink (#2251)
The goal (not yet achieved) is to mask the display geometry of Android devices and to make
the various devices identical for canvas and rect fingerprinting.
This patch is WIP, it is incomplete and contains visual glitch, do not activate.
This version contains points of interest marked in the code, which will be removed once the patch is complete.